Merkle test move only, no code changes
This moves the merkle integration tests into unit tests, to allow
removing the merkle code from the external interface.
